+++ controller/src/main/java/fr/insee/rem/controller/adapter/CensusAdapter.java
@@ -1,0 +1,13 @@
+package fr.insee.rem.controller.adapter;
+
+import fr.insee.rem.controller.sources.CensusSource;
+import fr.insee.rem.domain.dtos.SurveyUnitDto;
+import org.springframework.stereotype.Service;
+
+@Service
+public class CensusAdapter {
+    public SurveyUnitDto convert(CensusSource c) {
+
+        return SurveyUnitDto.builder().build();
+    }
+}