+++ controller/src/main/java/fr/insee/rem/controller/sources/CensusSource.java
@@ -1,0 +1,12 @@
+package fr.insee.rem.controller.sources;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class CensusSource {
+    private String nothing;
+}